Move all graphql to inner folder
