+++ src/redux/Topic/saga.js
@@ -4,7 +4,7 @@ import { TransactionType } from 'constants';
 
 import actions from './actions';
 import axios from '../../network/httpRequest';
-import { queryAllTopics, queryAllVotes } from '../../network/graphQuery';
+import { queryAllTopics, queryAllVotes } from '../../network/graphql/queries';
 import { satoshiToDecimal, processTopic } from '../../helpers/utility';
 import Routes from '../../network/routes';
