+++ src/stores/WalletStore.js
@@ -7,7 +7,7 @@ import { defineMessages } from 'react-intl';
 
 import axios from '../network/httpRequest';
 import Routes from '../network/routes';
-import { createTransferTx } from '../network/graphMutation';
+import { createTransferTx } from '../network/graphql/mutations';
 import { decimalToSatoshi } from '../helpers/utility';
 import Tracking from '../helpers/mixpanelUtil';
